- rename folders to spark_cde
