+++ dbt/adapters/spark_cde/__version__.py
