+++ dbt/adapters/spark_cde/impl.py
@@ -13,9 +13,9 @@ import dbt.exceptions
 from dbt.adapters.base import AdapterConfig
 from dbt.adapters.base.impl import catch_as_completed
 from dbt.adapters.sql import SQLAdapter
-from dbt.adapters.spark import SparkConnectionManager
-from dbt.adapters.spark import SparkRelation
-from dbt.adapters.spark import SparkColumn
+from dbt.adapters.spark_cde import SparkConnectionManager
+from dbt.adapters.spark_cde import SparkRelation
+from dbt.adapters.spark_cde import SparkColumn
 from dbt.adapters.base import BaseRelation
 from dbt.clients.agate_helper import DEFAULT_TYPE_TESTER
 from dbt.events import AdapterLogger
