+++ dbt/include/spark_cde/macros/adapters.sql
