+++ dbt/include/spark_cde/macros/materializations/snapshot.sql
