+++ dbt/include/spark_cde/macros/materializations/table.sql
