+++ dbt/include/spark_cde/macros/materializations/view.sql
