+++ dbt/include/spark_cde/macros/utils/assert_not_null.sql
