+++ dbt/include/spark_cde/macros/utils/bool_or.sql
