+++ dbt/include/spark_cde/macros/utils/concat.sql
