+++ dbt/include/spark_cde/macros/utils/datediff.sql
