+++ dbt/include/spark_cde/macros/utils/listagg.sql
