+++ dbt/include/spark_cde/macros/utils/split_part.sql
